change qei types to int16_t

+++ os/hal/ports/NRF5/LLD/hal_qei_lld.h
@@ -166,7 +166,7 @@ typedef enum {
 /**
  * @brief   QEI counter type.
  */
-typedef int32_t qeicnt_t;
+typedef int16_t qeicnt_t;
 
 /**
  * @brief   QEI delta type.
